added fader to camera prefab
