+++ Assets/Prefabs/Main Camera.prefab
@@ -16,6 +16,7 @@ GameObject:
   - 114: {fileID: 11411938}
   - 114: {fileID: 11444888}
   - 114: {fileID: 11487574}
+  - 114: {fileID: 11414556}
   m_Layer: 0
   m_Name: Main Camera
   m_TagString: MainCamera
@@ -109,6 +110,19 @@ MonoBehaviour:
   shaderFXAAPreset3: {fileID: 4800000, guid: c182fa94a5a0a4c02870641efcd38cd5, type: 3}
   shaderFXAAII: {fileID: 4800000, guid: cd5b323dcc592457790ff18b528f5e67, type: 3}
   shaderFXAAIII: {fileID: 4800000, guid: c547503fff0e8482ea5793727057041c, type: 3}
+--- !u!114 &11414556
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 142196}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 578558ca344f8474987b022beb281b20, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  fadeOutTexture: {fileID: 2800000, guid: cf6f92258193149e9871d6d98dae76f1, type: 3}
+  fadeSpeed: 1
 --- !u!114 &11444888
 MonoBehaviour:
   m_ObjectHideFlags: 1
